Change hawkbit interface to support ip addresses
Previously, hawkbit interface only supported a url/hostname and a port,
and internally it resolves to an IP address. This does not work for
network layers that rely on NAT64, like OpenThread.  Zephyr's
implementation of `getaddrinfo` is not aware of NAT64.  DNS will resolve
an IPV4 address that needs to be converted to IPV6 with the NAT64
prefix.

Instead, this commit alters the Hawkbit interface to allow providing an
IP address and hostname as a string, so an application can provide an
already resolved IP address if desired. This commit also changes the
usage of `hawkbit_runtime_config.server_addr` to point to an IP address,
and uses a new variable `hawkbit_runtime_config.server_hostname` for the
URL. If `server_addr` is not provided, hawkbit will resolve
`server_hostname`.

This commit also adds a check for hostname strings that are too long.
Previously, hawkbit would simply truncate the hostname and quietly fail
to connect. Now it returns an error to the application.